Locale/routes work as expected.
I have to admit that I don't understand how exactly the Rails 3/I18n
integration works nowadays. Apparently Rails always sets a :locale
param. So without that line in the config/i18n.rb initializer it would
always fill in the optional (:locale)/ path prefix segment (which is not
what we want). Also, even if I delete :locale from the params hash
Rails would *still* fill in the segment for generated URLs (whish is not
what we want).

This seems pretty odd to me. Rails shouldn't just add the :locale param
IMHO, but maybe I'm just confused, or maybe we're setting the param
somewhere else in the app and I've overseen it.

Anyhow this should now work. It would be interesting to understand why
Rails works this way though.